Calculate Functional Assessment of Chronic Illness Therapy - Fatigue (FACIT-F) sum score

Calculate FACIT-F sum score

Patients are asked "Please indicate how much each of the following statements has applied to you over the last 7 days by selecting the appropriate item."

calculate_facitf_sum(
  facitf_1,
  facitf_2,
  facitf_3,
  facitf_4,
  facitf_5,
  facitf_6,
  facitf_7,
  facitf_8,
  facitf_9,
  facitf_10,
  facitf_11,
  facitf_12,
  facitf_13
)

Arguments

facitf_1

vector for item "I feel fatigued."

facitf_2

vector for item "I feel weak all over."

facitf_3

vector for item "I feel listless ("washed out")."

facitf_4

vector for item "I feel tired."

facitf_5

vector for item "I have trouble starting things because I am tired."

facitf_6

vector for item "I have trouble finishing things because I am tired."

facitf_7

vector for item "I have energy."

facitf_8

vector for item "I am able to do my usual activities."

facitf_9

vector for item "I need to sleep during the day."

facitf_10

vector for item "I am too tired to eat."

facitf_11

vector for item "I need help doing my usual activities."

facitf_12

vector for item "I am frustrated by being too tired to do the things I want to do."

facitf_13

vector for item "I have to limit my social activity because I am tired."

Answers were coded /w levels "Not at all", "A little bit", "Somewhat", "Quite a bit" and "Very much"

Value

A numeric vector with sum score of FACIT-F