Calculate Patient Health Questionnaire Stress Scale (PHQ-Stress) sum score

Calculate sum score of PHQ-Stress

Patients are asked "During the past 4 weeks, how much have you been bothered by any of the following problems?"

calculate_phqs_sum(
  phqs_1,
  phqs_2,
  phqs_3,
  phqs_4,
  phqs_5,
  phqs_6,
  phqs_7,
  phqs_8,
  phqs_9,
  phqs_10
)

Arguments

phqs_1

vector for item "Worrying about your health"

phqs_2

vector for item "Your weight or how you look"

phqs_3

vector for item "Little or no sexual desire or pleasure during sex"

phqs_4

vector for item "Difficulties with husband/wife, partner/lover or boyfriend/girlfriend"

phqs_5

vector for item "The stress of taking care of children, parents, or other family members"

phqs_6

vector for item "Stress at wirk outside of the home or at school"

phqs_7

vector for item "Financial problems or worries"

phqs_8

vector for item "Having no one to turn to when you have a problem"

phqs_9

vector for item "Something bad that happened recently"

phqs_10

vector for item "Thinking or dreaming about something terrible that happened to you in the past - like your house being destroyed, a severe accident, being hit or assaulted, or being forced to commit a sexual act"

Answers were coded /w levels 0 = "Not bothered at all", 1 = "Bothered a little" and 2 = "Bothered a lot"

Value

A numeric vector with sum score of phqs