Read NAPKON Transferstelle export

Convenience wrapper for read_tsExport_raw, ' label_tsExport, factorize_tsExport and dates_tsExport.

read_tsExport(
  data_dir,
  separator = ";",
  decimal = ",",
  encoding = "UTF-8",
  quote_escape_detect = TRUE,
  escape_backslash = TRUE,
  labels = TRUE,
  factor = TRUE,
  dates = TRUE
)

Arguments

data_dir

string - location of the export

separator

Character. CSV file column separator.

decimal

Character.

encoding

Character. Encoding

quote_escape_detect

Boolean.Should the function try to detect the escaping method used in the data file? If TRUE, the reading process might be slower.

escape_backslash

Boolean.Are Quotes in strings escaped by a backslash? (The default with (german) secuTrial Exports).

labels

logical - add labels to variables and table

factor

logical - convert categorical variables to factor variables (ignored when reference values are not in a separate table)

dates

logical - convert date variables

Value

tsExportdata object - a list with one data.frame for each file on the export and a list containing the export options